requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
websockets>=12.0
